feat: added the API calls
Added multiple functionality: the get of all the books, the update when changing shelf and the seach of a book. However 2 bugs are open: on udating a book in the searchPage, the books are not updated, even if they appear correctly on the main page. Also removing a book from the shelves triggers an error, missing a remove from list function. Also the proptypes validation is still missing, as well as the readme file